requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
Django==4.1.1
djangorestframework==3.14.0
filelock==3.8.0
huggingface-hub==0.10.0
idna==3.4
numpy==1.23.3
packaging==21.3
pyparsing==3.0.9
pytz==2022.2.1
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
sqlparse==0.4.3
tokenizers==0.12.1
torch==1.12.1
tqdm==4.64.1
transformers==4.22.2
typing_extensions==4.3.0
tzdata==2022.4
urllib3==1.26.12
